+++ apps/astarte_data_updater_plant/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Do not accept invalid paths that have consecutive slashes.
+- Do not accept invalid paths in object aggregated interfaces.
 
 ## [0.10.1] - 2019-10-02
 ### Fixed